refactor: Create src/current-alias-redirects.ts
