+++ src/redirects.ts
@@ -1,4 +1,4 @@
-import { getPathInfo, Instance, isInstance, Url, CFEnvironment } from './utils'
+import { Instance, isInstance, Url, CFEnvironment } from './utils'
 
 const meetRedirects: Record<string, string | undefined> = {
   '/': 'vtk-ncrc-rdp',
@@ -24,7 +24,7 @@ const meetRedirects: Record<string, string | undefined> = {
   '/6': 'sui-yuwv-suh',
 }
 
-export async function redirects(request: Request, env: CFEnvironment) {
+export function redirects(request: Request, env: CFEnvironment) {
   const url = Url.fromRequest(request)
 
   if (url.subdomain === 'start') {
@@ -216,22 +216,6 @@ export async function redirects(request: Request, env: CFEnvironment) {
   ) {
     url.pathname = '/mathe-pruefungen'
     return url.toRedirect(302)
-  }
-
-  if (isInstance(url.subdomain)) {
-    const pathInfo = await getPathInfo(url.subdomain, url.pathname, env)
-
-    if (pathInfo !== null) {
-      const newUrl = new Url(url.href)
-      const { currentPath, instance, hash } = pathInfo
-
-      if (instance && isInstance(instance) && url.subdomain !== instance)
-        newUrl.subdomain = instance
-      if (url.pathname !== currentPath) newUrl.pathname = currentPath
-      if (hash !== undefined) newUrl.hash = hash
-
-      if (newUrl.href !== url.href) return newUrl.toRedirect(301)
-    }
   }
 
   if (
